requirements.txt
requests>=2.28.0
pdfplumber>=0.9.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
